Refactoring some stuff
