+++ plugins/ag-grid/src/js/src/AgGridPlugin.ts
@@ -2,7 +2,7 @@ import { type WidgetPlugin, PluginType } from '@deephaven/plugin';
 import { vsTable } from '@deephaven/icons';
 import type { dh } from '@deephaven/jsapi-types';
 import { LicenseManager } from '@ag-grid-enterprise/core';
-import { AgGridWidget } from './AgGridWidget';
+import { AgGridWidget } from './components';
 
 const key = import.meta.env?.VITE_AG_GRID_LICENSE_KEY ?? '';
 if (key != null && key !== '') {
